🔧 Moved directories to organize files better
