+++ forge/src/main/java/infinituum/fastconfigapi/forge/utils/ModScanner.java
@@ -1,8 +1,8 @@
 package infinituum.fastconfigapi.forge.utils;
 
 import infinituum.fastconfigapi.PlatformHelper;
-import infinituum.fastconfigapi.api.annotations.FastConfig;
 import infinituum.fastconfigapi.api.config.FastConfigFile;
+import infinituum.fastconfigapi.api.config.annotations.FastConfig;
 import infinituum.fastconfigapi.api.serializers.SerializerWrapper;
 import infinituum.fastconfigapi.api.utils.Global;
 import infinituum.fastconfigapi.api.utils.UnsafeLoader;
@@ -20,7 +20,7 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static infinituum.fastconfigapi.api.config.FastConfigs.MOD_ID;
+import static infinituum.fastconfigapi.api.FastConfigs.MOD_ID;
 
 public final class ModScanner {
     public static <T> Map<Class<T>, FastConfigFile<T>> getSidedConfigs(FastConfig.Side side) {
